refactor : station이 구간에 존재할 경우 delete 실패

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -7,17 +7,22 @@ import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
+import wooteco.subway.repository.SectionRepository;
 import wooteco.subway.repository.StationRepository;
 import wooteco.subway.utils.exception.NameDuplicatedException;
+import wooteco.subway.utils.exception.SubwayException;
 
 @Transactional
 @Service
 public class StationService {
 
     private final StationRepository stationRepository;
+    private final SectionRepository sectionRepository;
 
-    public StationService(StationRepository stationRepository) {
+    public StationService(StationRepository stationRepository,
+                          SectionRepository sectionRepository) {
         this.stationRepository = stationRepository;
+        this.sectionRepository = sectionRepository;
     }
 
     public StationResponse save(final StationRequest stationRequest) {
@@ -41,6 +46,9 @@ public class StationService {
     }
 
     public void deleteStation(final Long id) {
+        if (sectionRepository.isStationExist(id)) {
+            throw new SubwayException("[ERROR] 역이 구간에 존재하여 삭제할 수 없습니다.");
+        };
         stationRepository.deleteById(id);
     }
 }
